+++ v2/tools/release_manifest_test.go
@@ -22,8 +22,9 @@ type moduleEdit struct {
 		Path string
 	}
 	Require []struct {
-		Path    string
-		Version string
+		Path     string
+		Version  string
+		Indirect bool
 	}
 }